+++ apps/mobile/App.tsx
@@ -1,0 +1,18 @@
+/**
+ * Cohe Capital Mobile App
+ * Web3 Insurance DApp
+ */
+
+import React from 'react';
+import { StatusBar } from 'expo-status-bar';
+import AppProviders from './src/providers/AppProviders';
+import RootNavigator from './src/navigation/RootNavigator';
+
+export default function App() {
+  return (
+    <AppProviders>
+      <StatusBar style="light" />
+      <RootNavigator />
+    </AppProviders>
+  );
+}